pre-final

+++ src/main/java/com/increff/employee/pojo/orderitemPojo.java
@@ -39,7 +39,7 @@ public class orderitemPojo {
 	public double getPrice() {
 		return price;
 	}
-	public void setPrice(int price) {
+	public void setPrice(double price) {
 		this.price = price;
 	}
 	public int getQuantity() {
